init.sh

+++ init.sh
@@ -4,9 +4,9 @@
 
 
 # 初始化php的配置文件
-cp run.sh.example run.sh
-cp conf/php-fpm.conf.example conf/php-fpm.conf 
-cp conf/php.ini.example      conf/php.ini
+[ ! -f run.sh ]             && cp run.sh.example run.sh
+[ ! -f conf/php.ini ]       && cp conf/php.ini.example      conf/php.ini
+[ ! -f conf/php-fpm.conf ]  && cp conf/php-fpm.conf.example conf/php-fpm.conf 
 
 # 下载相关文件
 source ./download.sh
